// src/app/interfaces/sale-header.interface.ts
export interface SaleHeader
{
    id: number;
    ticketId: string;
    tipoVenta: string;
    canalDistribucion: string;
    centro: string;
    refClienteUno: string;
    refClienteDos: string;
    grupoVendedor: string;
    fecha: string;
    codigoEntrada: string;
    oficinaVenta: string;
    sociedad: string;
    anoEjercicio: string;
    condicionPagoId: string;
    metodoPagoId: string;
    tipoMoneda: string;
    monedaLocal: string;
    deudorId: string;
    guId: string;
    ventaDocId: string;
    transladoDocId: string;
    deliveryPickId: string;
    materialDocId: string;
    recepcionDocId: string;
    paqueteria: string;
    guiaRastreo: string;
    entregaDocId: string;
    facturaDocId: string;
    contableDocId: string;
    costoSemanal: string;
    costoMensua: string;
    totalMeses: string;
    aperturaId: string;
    origenVenta: string;
    startSale: string;
    endSale: string;
    processed: boolean;
}